Update[Info]: add timer color constants and moved getTimerColor function
