+++ src/constants/index.ts
@@ -1,1 +1,2 @@
 export * from "./winCombinations";
+export * from "./timerColors";
