+++ src/constants/timerColors.ts
@@ -1,0 +1,5 @@
+export const TIMER_COLORS = {
+  red: [154, 255],
+  green: [205, 0],
+  blue: [50, 0],
+};
